requirements.txt
streamlit>=1.30.0
google-generativeai==0.8.5
python-dotenv>=1.0.0
PyYAML>=6.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
